backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
requests==2.31.0
beautifulsoup4==4.12.3
